services/product-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Security (for JWT validation)
python-jose[cryptography]>=3.3.0

# HTTP Client
httpx>=0.25.0

# Redis
redis>=5.0.0

# Rate Limiting
slowapi>=0.1.9

# AWS S3
boto3>=1.29.0

# File handling
python-multipart>=0.0.6

# Logging
python-json-logger>=2.0.7

# Environment
python-dotenv>=1.0.0
